Separate CPU and CUDA implementations

HIP option?
